Add files via upload
The purpose of this code is to filter, visualize, and export a Modified Normalized Difference Water Index (MNDWI) image derived from a Landsat 8 satellite image using Google Earth Engine.

Here's the breakdown of the code:

Add the region of interest to the map: The Map.addLayer(Austin) and Map.centerObject(Austin, 10) commands add the Austin region of interest to the map and centers the map to this region at a zoom level of 10.

Search for Landsat 8 images: The code fetches Landsat 8 images for the Austin region within the 2022 calendar year, filtering only the images with cloud cover less than 1%.

Add the image collection to the map: The filtered image collection is then added to the map.

Load a specific Landsat 8 image: The code loads a specific Landsat 8 image (identified by 'LC08_027039_20221128') taken on November 28, 2022, and clips it to the Austin region.

Calculate MNDWI: The Modified Normalized Difference Water Index (MNDWI) is calculated using the green (band 3) and short-wave infrared (SWIR, band 6) bands of the selected image. MNDWI is an index that can be used for enhanced extraction of water features.

Visualize the MNDWI result: The MNDWI image is then added to the map with a color gradient palette. It ranges from green (for MNDWI values of -1), white (for MNDWI values of 0), and blue (for MNDWI values of 1).

Export the result: Finally, the MNDWI image is exported to your Google Drive in the "GEE_AUSTIN_MNDWI (2012-2022)" folder with the filename "Austin_LC08_027039_MNDWI_2022". The exported image will be in GeoTIFF format, a type of georeferenced raster image.