Add files via upload
This code is used to calculate and visualize the Normalized Difference Bare Soil Index (NDBSI) for a Landsat 8 image of Austin. It then exports the result to Google Drive. Here's what each section of the code does:

Setting up the map: This section sets the map to show the region of interest, which is Austin in this case. It adds Austin as a layer and centers the map on this location.

Searching for Landsat 8 images: This section sets up a variable landsat8sr as an ImageCollection object from Earth Engine's public data catalog. The function filterBounds() limits the collection to images that intersect Austin. filterDate() narrows down the collection to images taken within the year 2013, and filterMetadata() restricts the collection to images with a cloud cover of less than 5%.

Adding the image collection to the map: This part of the code adds the filtered image collection to the map and prints the collection to the console.

Loading a specific Landsat 8 image: The image variable represents a single Landsat 8 image identified by its specific ID ('LC08_027039_20131103'). This image is then clipped to the boundaries of the Austin region.

Calculating the NDBSI: This section calculates the Normalized Difference Bare Soil Index (NDBSI) using the formula (SWIR1 - SWIR2) / (SWIR1 + SWIR2). The image.expression() function is used to apply this formula to the swir1 and swir2 bands of the image.

Visualizing the NDBSI: This part of the code sets up the visualization parameters for the NDBSI result and adds the visualized NDBSI as a layer on the map.

Exporting the result: The Export.image.toDrive() function is used to save the NDBSI result to Google Drive. The exported image will be described as 'Austin_LC08_027039_NDBSI_2012' and saved in a folder called 'GEE_AUSTIN_NDBSI (2012-2022)'.

The NDBSI index is often used to identify bare soil areas, which can be useful for various applications like agriculture, urban planning, or land use and land cover mapping.