Add files via upload
The provided code is meant to calculate the Normalized Difference Moisture Index (NDMI) for a specific Landsat 8 image over the Austin region. The steps in the code are as follows:

Setting up the map: The code adds Austin as a layer on the map and centers the map over this region.

Searching for Landsat 8 images: The landsat8sr variable is assigned to a collection of Landsat 8 images. This collection is filtered to include only images that intersect with Austin, were captured between January 1, 2012, and December 30, 2013, and have a cloud cover of less than 1%.

Adding the image collection to the map: The filtered image collection is added to the map, and the collection is printed to the console for further inspection.

Loading a specific Landsat 8 image: The image variable represents a specific Landsat 8 image identified by its unique ID ('LC08_027039_20131103'). The image is then clipped to fit the boundaries of the Austin region.

Calculating NDMI: NDMI is calculated using the formula (NIR - SWIR) / (NIR + SWIR), where NIR is the near-infrared band and SWIR is the short-wave infrared band of the image. NDMI is a measure of the content of water in the observed area and can be useful for assessing vegetation health or tracking droughts.

Visualizing the NDMI: The NDMI results are visualized on the map, with a color palette ranging from blue (for -1 values) to green (for 1 values).

Exporting the result: The Export.image.toDrive() function is used to export the NDMI result to Google Drive. The exported image will have the description 'Austin_LC08_027039_NDMI_2013' and will be saved in a folder named 'GEE_AUSTIN_NDMI (2012-2022)'.