Add files via upload
This code is a script that uses the arcpy module to perform spatial analysis on raster data. The script imports the necessary modules, sets the input and output folders, and defines the paths to four shapefiles. The script then checks out the ArcGIS Spatial Analyst extension and sets the overwriteOutput parameter to True.

The script then loops through a range of years from 1992 to 2022 and performs the following steps for each year:

Sets the input file path based on the year and whether it is before or after 2012.
Sets the output file path based on the year and the output folder.
Uses the ExtractByMask function from the arcpy.sa module to clip the input raster using one of the four shapefiles.
Saves the clipped raster to the specified output file.
Prints a message indicating that the clipping was successful.
The script performs these steps for each of the four shapefiles, resulting in four sets of clipped rasters. Finally, the script checks in the ArcGIS Spatial Analyst extension. This code is used to clip raster data using shapefiles and save the resulting clipped rasters to specified output folders.